requirements.txt
Flask==3.0.0
pymongo==4.6.3
python-dotenv==0.21.0
bcrypt==3.2.0
